dataloader and mask nits
